[INJIMOB-1655]: update folder structure to make it more organised and compact
