+++ vc-verifier/kotlin/vcverifier/src/main/java/io/mosip/vercred/vcverifier/CredentialVerifier.kt
@@ -5,7 +5,7 @@ import io.mosip.vercred.vcverifier.constants.CredentialFormat
 import io.mosip.vercred.vcverifier.constants.CredentialFormat.LDP_VC
 import io.mosip.vercred.vcverifier.constants.CredentialVerifierConstants.EXCEPTION_DURING_VERIFICATION
 import io.mosip.vercred.vcverifier.constants.CredentialVerifierConstants.VERIFICATION_FAILED
-import io.mosip.vercred.vcverifier.credentialverifier.verifiablecredential.CredentialVerifierFactory
+import io.mosip.vercred.vcverifier.credentialverifier.CredentialVerifierFactory
 import io.mosip.vercred.vcverifier.data.VerificationResult
